feat(loop tick 50): the engine runs in a browser — wasm feasibility proven, and one 32-bit bug
The in-browser demo directive's load-bearing unknown is "does the engine even compile to
wasm32-unknown-unknown?" Everything downstream (curated snapshots, canvas, side-by-side vs
Chromium, GitHub Pages) is scaffolding around that one fact — so probe it FIRST, before building
anything, exactly as the WPT runner probed testharness.js.

RESULT: the entire render pipeline minus JS compiles to wasm. dom, css+STYLO, layout(taffy),
paint(tiny-skia), html, text — all of it, proven repeatably by ./scripts/wasm-check.sh. Stylo
was the real risk and it cleared. SpiderMonkey (C++) is deliberately absent, which is also WHY
the demo is JS-free. The demo is feasible.

THE ONE GENUINE BLOCKER, found by the probe and fixed: NodeId packed generation<<32 | index
into a USIZE — and wasm32's usize is 32 BITS, so the shift OVERFLOWED and manuk-dom did not even
compile ("this arithmetic operation will overflow"). This is the exact class of bug a probe
exists to find: invisible on the 64-bit dev machine, fatal on the target. The fix is a u64
backing — byte-identical to usize on 64-bit, correct on 32-bit — so the arena is now
POINTER-WIDTH-INDEPENDENT, which also hardens the ARM/cross-platform target the process-model
directive named. Any type that packs bit-fields into a pointer-width integer is a latent 32-bit
bug.

The ripple was ~20 sites (NodeId(x as usize) -> as u64, a couple of .map(NodeId) over usize
ranges, one shell node.0), all mechanical, and NATIVE IS UNREGRESSED — the full wall is green.
G_ARENA_U64 pins the packing semantics so a future "simplify NodeId back to usize" cannot
silently reintroduce the overflow, and it is proven falsifiable.

THE PLAN is written (docs/loop/DEMO.md): real (Stylo/Taffy/tiny-skia executing in the visitor's
browser; live scroll/click/hover and CSS state), not real and SAID SO IN-PRODUCT (no JS, no
arbitrary fetch — curated snapshots re-derived from the oracle's cluster registry), single-
threaded by choice (keeps GitHub Pages clean: no COOP/COEP, no SharedArrayBuffer), a
side-by-side-vs-Chromium toggle so realness is CHECKABLE not assertable, its own non-blocking CI
lane on a stable branch, and the maintenance rule that the demo path's output must not diverge
from the native path's.

Directly serves the standing concern: the same codebase must work for the wasm demo, native
x86-64, AND mac/windows/linux without regressing capability. The u64 fix advances all of them at
once: 32-bit safety for wasm, pointer-width independence for ARM, zero native regression.

Next (the demo is feasible, not yet built): the demo/ wasm crate wiring the already-proven
pipeline to a canvas, the snapshot-baking build step, the JS-glue shell, and the GitHub Pages lane.

TICK SHAPE: infrastructure
[no-pattern]

WIKI: docs/wiki/build-and-dependencies.md — the render pipeline compiles to wasm32 (Stylo included) but a
usize in a packed handle does not: NodeId's generation<<32|index overflowed wasm32's 32-bit usize; back
bit-packed handles with an explicit u64, never usize, which serves wasm and ARM at once.

+++ agent/src/automation.rs
@@ -177,7 +177,7 @@ mod tests {
 
     fn n(id: u64, role: Role, name: &str, boxed: bool, children: Vec<A11yNode>) -> A11yNode {
         A11yNode {
-            node: NodeId(id as usize),
+            node: NodeId(id as u64),
             role,
             name: name.to_string(),
             bbox: boxed.then(|| Rect {
