+++ agent/src/targeting.rs
@@ -219,7 +219,7 @@ mod tests {
 
     fn n(id: u64, role: Role, name: &str, bbox: Option<Rect>, children: Vec<A11yNode>) -> A11yNode {
         A11yNode {
-            node: NodeId(id as usize),
+            node: NodeId(id as u64),
             role,
             name: name.to_string(),
             bbox,
@@ -356,7 +356,7 @@ mod tests {
         let t = resolve_target(&tree, "sign in", viewport).expect("a target");
         assert_eq!(
             t.node,
-            NodeId(4 as usize),
+            NodeId(4 as u64),
             "the prominent nav button, not the footer link"
         );
         assert_eq!(t.name, "Sign in");
